added generalised vector rotation functionality
Added generalised vector rotations that should work for oblique incidence angles as well, will test those tomorrow. On the way was able to add a feature that allows the user to choose the preferred scattering angle that will make verification of results easier. The added cross products add alot to the complexity and thus are making utkarsh's PC lag. Will have to be careful when increasing photon counts.